+++ azure/storage/_constants.py
@@ -15,7 +15,7 @@
 import platform
 
 __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
-__version__ = '0.34.2'
+__version__ = '0.34.3'
 
 # x-ms-version for storage service.
 X_MS_VERSION = '2016-05-31'
@@ -36,8 +36,8 @@ DEV_TABLE_HOST = '127.0.0.1:10002'
 DEV_ACCOUNT_NAME = 'devstoreaccount1'
 DEV_ACCOUNT_KEY = 'Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw=='
 
-# Socket timeout in seconds is 11
-SOCKET_TIMEOUT = 11
+# Socket timeout in seconds
+DEFAULT_SOCKET_TIMEOUT = 20
 
 #Encryption constants
 _ENCRYPTION_PROTOCOL_V1 = '1.0'