+++ azure/storage/file/fileservice.py
@@ -124,7 +124,7 @@ class FileService(StorageClient):
 
     def __init__(self, account_name=None, account_key=None, sas_token=None, 
                  protocol=DEFAULT_PROTOCOL, endpoint_suffix=SERVICE_HOST_BASE, 
-                 request_session=None, connection_string=None):
+                 request_session=None, connection_string=None, socket_timeout=None):
         '''
         :param str account_name:
             The storage account name. This is used to authenticate requests 
@@ -149,6 +149,9 @@ class FileService(StorageClient):
             request session. See
             http://azure.microsoft.com/en-us/documentation/articles/storage-configure-connection-string/
             for the connection string format.
+        :param int socket_timeout:
+            If specified, this will override the default socket timeout. The timeout specified is in seconds.
+            See DEFAULT_SOCKET_TIMEOUT in _constants.py for the default value.
         '''
         service_params = _ServiceParameters.get_service_parameters(
             'file',
@@ -158,7 +161,8 @@ class FileService(StorageClient):
             protocol=protocol, 
             endpoint_suffix=endpoint_suffix,
             request_session=request_session,
-            connection_string=connection_string)
+            connection_string=connection_string,
+            socket_timeout=socket_timeout)
             
         super(FileService, self).__init__(service_params)