+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
+## Version 0.34.3:
+- All: Made the socket timeout configurable. Increased the default socket timeout to 20 seconds.
+- All: Fixed a bug where SAS tokens were being duplicated on retries
+
 ## Version 0.34.2:
 
 ### All: